移除多数test文件夹下的程序，更改目录结构

+++ ASCII.c
@@ -13,10 +13,10 @@
 int main() {
 	printf("Num |  Hex  |    Bit    | Char\n");
 	for (unsigned short a = 0; a < 260; a++) {
-		unsigned int out = 0;
+		unsigned long out = 0;
 		for (int i = 0; i < 32; i++) out = out * 10 + ((a << i) >> 31);
 		out = 0 - out;
-		printf("%03d | 0x%03x | %09d | ", a, a, out);  //循环打印输出
+		printf("%03d | 0x%03x | %09ld | ", a, a, out);  //循环打印输出
 		switch (a) {
 		case '\r':
 			printf("\\r\n");
